Use FALLBACK when an invalid language is provided
Rather than error()'ing and exiting the code stack, it's probably better to just print the error and continue on using the FALLBACK language. This should be the least disruptive to the end-user (they'll still potentially be able to see a translated phrase), and is more 3rd-party friendly by allowing their code to keep executing.